Posts can now be password protected
To enable add a non false password metadata field.
The password will be added as the password metadata value.

An optional sub directory of posts is automatically added on `htmd start`.
This new folder can be ignored when using source control.
Otherwise the password would be compromised
and so would the post contents which are supposed to be secret.

When the page is loaded the encrypted post will be in JavaScript and
a password prompt will open.
The post will be decrypted with the provided value.